nfacct: adding quota capabilities
The accounting framework now supports quota at the packet and byte
level.  The tool is simply enhanced with two optional arguments to
specify whether accounting is for byte or packet and the limit
associated with each.  Also adding a monitor mode that listens for
quota attainment notifications.

Examples:

/* create an accounting object that isn't associated to a quota */
$ nfacct add first_no_quota

/* create a quota object with byte count limited to 50 byte */
$ nfacct add second_quota byte 50

/* create a quota object with packet count limited to 5 */
$ nfacct add third_quota packet 5

From there the accounting objects can be used in iptables the same
way as they did before:

/* limit the number of icmp packets allowed through the OUTPUT chain */
$ iptables -I OUTPUT -p icmp -m nfacct --nfacct-name third_quota --jump REJECT

/* listening for quota attainment notification */
$ nfacct monitor

Everything else works the same way.

Change-Id: I98532278953f51d8c8a65926510f5a469361ab0f
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>